pyproject.toml
[project]
name = "softdesk"
version = "0.1.0"
description = ""
authors = [
    {name = "Michael",email = "56340359+SynysterRev@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "drf-nested-routers (>=0.94.1,<0.95.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
